server/requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
python-docx
requests
httpx
python-jose[cryptography]
passlib[bcrypt]